requirements.txt
faiss-cpu
sentence-transformers
requests
python-dotenv
numpy